test: split applySourceDocuments.test.ts by topic
At 1,786 lines this was the largest remaining test file, with 3 top-level
describes and no shared setup between them - each test carries its own
inline fixtures, so the split has zero cross-file coupling. Splits into
test/csm/applySourceDocuments/{complexQueries,simpleQueries,handlingPerspectives}.test.ts
along those existing describe boundaries.

The resulting files are uneven (163/255/1372 lines) rather than the roughly
equal thirds a naive split would suggest, because the fixture-heavy test data
is concentrated in the "handling perspectives" tests (3 tests, 12 inline
snapshots) versus one test each in the other two topics. Splitting still
separates 3 genuinely independent topics into their own files rather than one
mixed-concern file; the size skew is inherent to the content, not the split.

Pure move: 5 tests before and after, same assertions, same test names.
